+++ src/input/fromFilesystem.ts
@@ -1,3 +1,4 @@
+import binaryData from 'binary-extensions';
 import commonPrefix from 'common-path-prefix';
 import ignore from 'ignore';
 import FS from 'node:fs';
@@ -7,10 +8,6 @@ import { getFileExtension, normPath } from '../program/fs/normalisedPath.js';
 import walkTree from '../program/fs/walkTree.js';
 import parseGitattributes from '../program/parsing/parseGitattributes.js';
 import * as T from '../types/types.js';
-
-const binaryData = JSON.parse(
-	FS.readFileSync(new URL('../../node_modules/binary-extensions/binary-extensions.json', import.meta.url), 'utf-8')
-) as string[];
 
 type FileInput = string;